+++ src/mixins/QuestionMixin.js
@@ -26,6 +26,7 @@ import axios from '@nextcloud/axios'
 import debounce from 'debounce'
 
 import logger from '../utils/Logger.js'
+import GenRandomId from '../utils/GenRandomId.js'
 import OcsResponse2Data from '../utils/OcsResponse2Data.js'
 import Question from '../components/Questions/Question.vue'
 
@@ -177,6 +178,14 @@ export default {
 		 * Can question be moved down in order?
 		 */
 		canMoveDown: {
+			type: Boolean,
+			default: false,
+		},
+
+		/**
+		 * isComposing for IME handling
+		 */
+		isIMEComposing: {
 			type: Boolean,
 			default: false,
 		},
@@ -408,6 +417,64 @@ export default {
 				this.focusIndex(options.length - 1)
 			})
 			this.isLoading = false
+		},
+
+		/**
+		 * Add a new empty answer locally
+		 * @param {InputEvent} event The input event that triggered adding a new entry
+		 */
+		addNewEntry({ target, isComposing }) {
+			/*
+			 * Check for !isComposing needed for languages using IME like Japanese or Chinese
+			 * Check for !this.isComposing needed for IME inputs handled by CompositionEvents
+			 * Check for target.value !== '' needed for Linux/Mac for characters like á or è
+			 */
+			if (!isComposing && !this.isIMEComposing && target.value !== '') {
+				// Add local entry
+				const options = [
+					...this.options,
+					{
+						id: GenRandomId(),
+						questionId: this.id,
+						text: target.value,
+						local: true,
+					},
+				]
+
+				// Reset the "new answer" input if needed
+				if (this.$refs.pseudoInput) {
+					this.$refs.pseudoInput.value = ''
+				}
+
+				// Update questions
+				this.updateOptions(options)
+
+				this.$nextTick(() => {
+					// Set focus to the created input element
+					this.focusIndex(options.length - 1)
+
+					// Trigger onInput on new AnswerInput for posting the new option to the API
+					this.$refs.input[options.length - 1].onInput()
+				})
+			}
+		},
+
+		/**
+		 * Handle compostion start event for IME inputs
+		 */
+		onCompositionStart() {
+			this.isIMEComposing = true
+		},
+
+		/**
+		 * Handle compostion end event for IME inputs
+		 * @param {CompositionEvent} event The input event that triggered adding a new entry
+		 */
+		onCompositionEnd({ target, isComposing }) {
+			this.isIMEComposing = false
+			if (!isComposing) {
+				this.addNewEntry({ target, isComposing })
+			}
 		},
 	},
 }
